Added support for BANZAI images to run through mtdlink: added extract_sci_image() to make a new fits file with only the image from the BANZAI multi-extension fits file to put into the working directory of pipeline_astrometry.py; this removes the symbolic links in the working directory for BANZAI fits files.

+++ neoexchange/photometrics/external_codes.py
@@ -304,7 +304,7 @@ def run_scamp(source_dir, dest_dir, fits_catalog_path, binary=None, dbg=False):
     return retcode_or_cmdline
 
 @timeit
-def run_mtdlink(source_dir, dest_dir, fits_file_list, num_fits_files, param_file, pa_rate_dict, binary=None, catalog_type='ASCII', dbg=False):
+def run_mtdlink(source_dir, dest_dir, fits_file_list, num_fits_files, param_file, pa_rate_dict, catfile_type, binary=None, catalog_type='ASCII', dbg=False):
     '''Run MTDLINK (using either the binary specified by [binary] or by
     looking for 'mtdlink' in the PATH) on the passed <fits_files> with the results
     and any temporary files created in <dest_dir>. <source_dir> is the path
@@ -330,11 +330,12 @@ def run_mtdlink(source_dir, dest_dir, fits_file_list, num_fits_files, param_file
         fits_file = os.path.basename(f)
         if fits_file != f:
             fits_file = os.path.join(dest_dir, fits_file)
-            # If the file exists and is a link (or a broken link), then remove it
-            if os.path.lexists(fits_file) and os.path.islink(fits_file):
-                os.unlink(fits_file)
-            if not os.path.exists(fits_file):
-                os.symlink(f, fits_file)
+            if not 'BANZAI' in catfile_type:
+                # If the file exists and is a link (or a broken link), then remove it
+                if os.path.lexists(fits_file) and os.path.islink(fits_file):
+                    os.unlink(fits_file)
+                if not os.path.exists(fits_file):
+                    os.symlink(f, fits_file)
         symlink_fits_files.append(fits_file)
 
     linked_fits_files = ' '.join(symlink_fits_files)
